Israel Defense Forces
A set of IDF loadouts in olive drab with the IWI Tavor (CTAR-21 variant) as the primary rifle. I used the vanilla Negev as the LMG, while the AT riflemen carry the LAW while the MAT teams carry the RPG-7. The pilots carry a PP-2000 as a stand-in for the Uzi/MAC-10.